requirements.txt
pytesseract
Pillow
pdf2image
Flask
Flask-CORS
Werkzeug